Assign commits to the standard GitHub Actions bot
The action and reusable workflow commits the generated documentation to the repository. A username and email address
must be configured for this commit.

Previously, an arbitrary GitHub account was used for this purpose. It is questionable whether it is appropriate to
attribute commits to an arbitrary user who didn't make them and which was not explicitly provided for such a purpose. In
addition, doing so is problematic because we require all contributors to sign a CLA, and so we would need to set up an
exemption for this arbitrary user in order to allow the commits to pass the CLA signing check.

So the better approach is to assign the commit to the standard GitHub Actions bot account. This account is already
assigned machine generated commits by the infrastructure of several of Arduino's repositories and is already exempted
from the CLA check.

An alternative would have been to use Arduino's general purpose "ArduinoBot" machine account. This account is assigned
automated commits by some infrastructure used exclusively in Arduino's own projects. However, the
"arduino/render-docs-github-action" action and reusable workflow may also be used in projects not owned by Arduino, and
it probably isn't appropriate to assign commits to ArduinoBot in those repositories.