Make commit author configurable
The users of the action and reusable workflow may wish to attribute commits for the rendered documentation to a specific
user. This can now be done via the newly added `committer-name` and `committer-email` action inputs.

The inputs are optional, using the GitHub Actions bot account as the default author, as was done before the addition of
the inputs.